auth: Revamp authentication and authorization stack
Authorization is different from authentication, but Ktor doesn't have
special support for it. Create an AuthorizationResult enum instead and
set it as an ApplicationCall attribute to provide more detiled error
info to users.

This also makes authorization more robust and more organized as it has
been split into a separate package.