content: Serve static content at server root
Add the ability to serve static content at the server root, protected by
the same authentication wall used for exclusive content under /exclusive.
This makes it convenient to host frontends that use exclusive content
without having to deal with authentication or redirecting.

Note that content filters are *not* applied to static content.