content: exclusive: Add support for non-patron grants
This adds the ability for the creator to generate grant links for
non-patrons that should have access to exclusive content.

Grants are JSON payloads bound to the file path and time. In grant
links, grants are encrypted with AES-256-GCM for authentication and
security, and encoded in hexadecimal form for URL validity. They also
contain a tag that can be used by content filters.