stripe: Remove failed purchase info from database
If the email fails to send (and thus the purchase fulfillment webhook
fails), it's better to remove the info and start over when Stripe
retries the webhook (or the creator manually resends the webhook from
the dashboard). Otherwise, further attempts to fulfill the purchase will
succeed but do nothing because of the idempotency key, so the purchase
never actually gets fulfilled.