patreon: Fix HTTP requests with new Apache httpcomponents
- Ktor 2.0 brought in a newer version of Apache httpcomponents as a
  transitive dependency, like GeoIP did before

Fix it properly:

Newer versions of Apache httpcomponents result in pathSuffix being "/current_user" with
Patreon's usage. The path is concatenated to the base API URL, resulting in a double slash:
https://www.patreon.com/api/oauth2/api//current_user?include=pledges

Patreon API responds with a 308 redirect + HTML page to remove the double slash. Java HTTP
client doesn't follow redirects, so we need to fix the URL manually.

Use reflection to inject the fixed RequestUtil.